need to fix CGI

+++ includes/webserv.hpp
@@ -60,7 +60,7 @@ struct location
 	std::string index_html;
 	std::string upload_directory;
 	char allowed_method[4];
-	int client_body_size;
+	unsigned long client_body_size;
 	bool directory_listing;
 	std::map<int, std::string> error_pages;
 	location &operator=(location &l);
@@ -78,7 +78,7 @@ struct config
 	std::map<int, std::string> error_pages;
 	char allowed_method[4];
 	int port;
-	int client_body_size;
+	unsigned long client_body_size;
 	bool directory_listing;
 	std::map<std::string, location> locations; // PATH -> location
 	config &operator=(config &c);
